split camera into QRFragment and cameraFragment. Just need to fix up the camera implementation for storing the image. QR code works and I'm looking to handle the marketing and sign up QR codes.
