--- app/src/main/java/com/example/genzgpt/Controller/CameraActivity.java
@@ -1,127 +1,0 @@
-package com.example.genzgpt.Controller;
-import android.Manifest;
-import android.app.Activity;
-import android.content.Intent;
-import android.content.pm.PackageManager;
-import android.net.Uri;
-import android.os.Bundle;
-import android.provider.MediaStore;
-import android.util.Log;
-
-import androidx.core.app.ActivityCompat;
-import androidx.core.content.ContextCompat;
-
-import com.google.zxing.integration.android.IntentIntegrator;
-import com.google.zxing.integration.android.IntentResult;
-
-import java.io.File;
-import java.io.IOException;
-
-public class CameraActivity extends Activity {
-    private static final int REQUEST_IMAGE_CAPTURE = 1;
-    private static final int REQUEST_CAMERA_PERMISSION = 200;
-    private Uri photoURI;
-    private String qrCodeResult;
-    private boolean isQrCode;
-
-    public CameraActivity(boolean isQrCode) {
-        this.isQrCode = isQrCode;
-        if (checkPermissions()) {
-            if (isQrCode) {
-                initiateQrScan();
-            } else {
-                initiateImageCapture();
-            }
-        } else {
-            requestPermissions();
-        }
-    }
-
-    private boolean checkPermissions() {
-        return ContextCompat.checkSelfPermission(this, Manifest.permission.CAMERA) == PackageManager.PERMISSION_GRANTED
-                && ContextCompat.checkSelfPermission(this, Manifest.permission.WRITE_EXTERNAL_STORAGE) == PackageManager.PERMISSION_GRANTED;
-    }
-
-    private void requestPermissions() {
-        ActivityCompat.requestPermissions(this, new String[]{Manifest.permission.CAMERA, Manifest.permission.WRITE_EXTERNAL_STORAGE}, REQUEST_CAMERA_PERMISSION);
-    }
-
-    @Override
-    public void onRequestPermissionsResult(int requestCode, String[] permissions, int[] grantResults) {
-        if (requestCode == REQUEST_CAMERA_PERMISSION) {
-            if (grantResults.length > 0 && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
-                if (isQrCode) {
-                    initiateQrScan();
-                } else {
-                    initiateImageCapture();
-                }
-            } else {
-                Log.e("CameraActivity", "Permission denied");
-            }
-        }
-    }
-
-    private void initiateQrScan() {
-        IntentIntegrator integrator = new IntentIntegrator(this);
-        integrator.setOrientationLocked(false);
-        integrator.initiateScan();
-    }
-
-    private void initiateImageCapture() {
-        Intent takePictureIntent = new Intent(MediaStore.ACTION_IMAGE_CAPTURE);
-        if (takePictureIntent.resolveActivity(getPackageManager()) != null) {
-            File photoFile = null;
-            try {
-                photoFile = createImageFile();
-            } catch (IOException ex) {
-                Log.e("CameraActivity", "Error occurred while creating the File", ex);
-            }
-            if (photoFile != null) {
-                photoURI = Uri.fromFile(photoFile);
-                takePictureIntent.putExtra(MediaStore.EXTRA_OUTPUT, photoURI);
-                startActivityForResult(takePictureIntent, REQUEST_IMAGE_CAPTURE);
-            }
-        }
-    }
-
-    private File createImageFile() throws IOException {
-        // Create an image file name
-        String imageFileName = "JPEG_" + System.currentTimeMillis() + "_";
-        File storageDir = getExternalFilesDir(null);
-        File image = File.createTempFile(
-                imageFileName,  /* prefix */
-                ".jpg",         /* suffix */
-                storageDir      /* directory */
-        );
-        return image;
-    }
-
-    @Override
-    protected void onActivityResult(int requestCode, int resultCode, Intent data) {
-        if (requestCode == REQUEST_IMAGE_CAPTURE && resultCode == RESULT_OK) {
-            // The image is saved to the file defined by photoURI
-            // You can use this URI elsewhere in your application
-        } else {
-            IntentResult result = IntentIntegrator.parseActivityResult(requestCode, resultCode, data);
-            if (result != null) {
-                if (result.getContents() == null) {
-                    Log.d("CameraActivity", "Cancelled scan");
-                } else {
-                    qrCodeResult = result.getContents();
-                }
-            } else {
-                super.onActivityResult(requestCode, resultCode, data);
-            }
-        }
-    }
-
-    // Getter method for the photoURI
-    public Uri getPhotoURI() {
-        return photoURI;
-    }
-
-    // Getter method for the QR code result
-    public String getQrCodeResult() {
-        return qrCodeResult;
-    }
-}
