+++ app/src/main/java/com/example/genzgpt/View/UserProfileFragment.java
@@ -1,4 +1,4 @@
-package com.example.genzgpt;
+package com.example.genzgpt.View;
 
 import android.os.Bundle;
 
@@ -7,6 +7,8 @@ import androidx.fragment.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import com.example.genzgpt.R;
 
 /**
  * A simple {@link Fragment} subclass.
