+++ app/src/main/res/layout/attendee_list_activity.xml
@@ -38,7 +38,7 @@
         android:layout_marginTop="4dp"
         android:layout_marginEnd="8dp"
         android:layout_toStartOf="@+id/number_of_attendees"
-        android:text="@string/total"
+        android:text="@string/number"
         android:textSize="20sp" />
 
     <TextView
@@ -49,7 +49,7 @@
         android:layout_alignParentEnd="true"
         android:layout_marginTop="4dp"
         android:layout_marginEnd="2dp"
-        android:text="@string/number"
+        android:text="@string/total"
         android:textSize="20sp" />
 
 </RelativeLayout>
