+++ app/src/main/res/layout/fragment_event_host.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".EventHostFragment">
+    tools:context=".View.EventHostFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
