+++ app/src/main/res/layout/fragment_my_events.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MyEventsFragment">
+    tools:context=".View.MyEventsFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
