+++ app/src/main/res/layout/fragment_user_profile.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".UserProfileFragment">
+    tools:context=".View.UserProfileFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
