+++ app/src/main/res/layout/request_permissions.xml
@@ -1,0 +1,22 @@
+<?xml version="1.0" encoding="utf-8"?>
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:padding="16dp">
+
+    <TextView
+        android:id="@+id/permissions_explanation"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="@string/this_app_requires_camera_and_storage_permissions_to_function_properly"
+        android:textSize="16sp" />
+
+    <Button
+        android:id="@+id/request_permissions_button"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_below="@id/permissions_explanation"
+        android:layout_marginTop="24dp"
+        android:text="@string/request_permissions" />
+
+</RelativeLayout>
